added notes and minor refactoring to send hardware id to all calls
