+++ .github/workflows/draftnewrelease.yml
@@ -22,7 +22,7 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # This token is provided by Actions, you do not need to create your own token
       with:
         tag_name: ${{ github.ref }}
-        release_name: Version 3.2.0.0
+        release_name: Version 4.0.0.0
         body: |
           - First Change
           - Second Change
